sweetened search, ingest, var handling, and users
A plethora of updates here. Moved search and ingest functions to
strictly identical calls between CDM and razuna and put the functions in
imree_functions and included them globally in the config.php and
config_setup.php. The repositories you can access is now controlled by
the groups a user is assigned to. Those values assignments are stored in
sql. The new 'sources' table controls the specific function name that
calls the search and ingest (those the need for those to be abstracted
and consistantly parameterized.